+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
 ## 1.0 (unreleased)
-- auto-importing works in the `--setup` clause 
+- auto-importing works in the `--setup` clause
+- will not internally change commands starting with a reserved keyword (ex: `if s ==` will not be changed to `s = if s ===`)
 
 ## 0.9 (2020-12-02)
 - other modules added for auto-import