+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0 (unreleased)
 - auto-importing works in the `--setup` clause
 - will not internally change commands starting with a reserved keyword (ex: `if s ==` will not be changed to `s = if s ===`)
+- CHANGED: `--end` clause renamed from `--finally`
 
 ## 0.9 (2020-12-02)
 - other modules added for auto-import